rename timeStampFormatter.ts to utils.ts
